refactoring some things into common kernels to reduce need for any new code in diff ik.
