+++ newton/_src/controllers/impl/operational_space/model_free.py
@@ -115,24 +115,24 @@ from ...controller import ControllerBase
 from ...utils import _validate_array
 from .._common import (
     _add_term_kernel,
+    _apply_spatial_matrix_kernel,
     _block_matrix_vector_multiply_kernel,
+    _invert_spd_block_kernel,
+    _null_space_projector_kernel,
     _pd_term_kernel,
+    _pose_error_kernel,
     _read_port,
     _scatter_port_kernel,
+    _task_matrix_times_jacobian_kernel,
 )
 from ._common import (
     _apply_generalized_task_specification_matrix_kernel,
     _apply_mass_matrix_inv_on_right_kernel,
-    _apply_spatial_matrix_kernel,
-    _invert_spd_block_kernel,
     _jacobian_times_jacobian_transpose_kernel,
     _jacobian_transpose_force_kernel,
-    _null_space_projector_kernel,
     _operational_space_mass_matrix_inverse_kernel,
-    _pose_error_kernel,
     _pose_twist_to_frame_kernel,
     _rotate_jacobian_to_frame_kernel,
-    _task_matrix_times_jacobian_kernel,
     _task_space_pd_kernel,
     _wrench_feedback_kernel,
     _wrench_feedforward_kernel,
